requirements.txt
Django>=5.2.7,<5.3
djangorestframework>=3.16.1,<3.17
psycopg2-binary>=2.9.11,<2.10
djangorestframework-simplejwt>=5.5.1,<5.6
celery>=5.5.3,<5.6
redis>=7.0.0,<7.1